Fix Firefox problem due to use of jQuery.
The problem is that a jQuery query result is being passed directly as
nodes to the D3 layout.  D3 writes various properties, such as
"children", directly to each node.  This can cause issues if there is an
existing property with the same name being used for another purpose,
such as jQuery's "children" in this case.

See also:

  http://stackoverflow.com/questions/14967381/using-xml-as-a-datasource-in-d3js-layout-doesnt-work-in-firefox